Made BasicFlatValue and Value's constructors explicit.
This is to prevent ambiguity for operation overloads. List's constructor now
automatically converts its elements to Value, to reduce verbosity. Set's
constructor does similarly, taking any number of Mappings, which automatically
convert their keys and values to the necessary type. However this means that for
each mapping argument, it is necessary to explicitly construct a Mapping:
Mapping{KEY, VALUE}.