Evaluating Expressions mostly implemented. Added a unary operator.
The '/' operator when used like '/ident' evaluates to a String of the variable
name: 'ident'. This removes issues with ambiguity for 'x:y', for a Map x. Now
'x:y' refers to an index which is the value contained by 'y', where 'x:/y'
refers to an index with a string value 'y'.

Most operators are now implemented, though some operators need some more work to
function as intended.